Split notification-settings.tsx into a hook and section components
notification-settings.tsx was 646 lines mixing all notification state
and logic with the full render tree. Extracted:

- use-notification-settings.ts — all state (15 useState), refs, effects,
  and handlers (save/test/toggle/permission) as a single logic hook
- notification-settings-constants.ts — NotifyEventType, response type,
  EVENT_OPTIONS
- notification-device-sections.tsx — SoundCuesSection + TipsSection
  (device-only local toggles)
- browser-notifications-section.tsx — Browser Notifications region
  (owns its platform detection)
- slack-notifications-section.tsx — Slack webhook + event toggles +
  save/test actions

NotificationSettings is now a 75-line composition shell. Pure structural
refactor — DOM, classes, testids, and behavior are unchanged.
